fix: widen downscale_to area count to u64 (resize.rs)
The per-destination-pixel source-area count is the divisor used to
average each destination pixel. It was computed with a u32 multiply while
the sum accumulator was already u64, so for large-source / small-dest
ratios the product overflowed u32: debug panic, or release wrap to 0
(black pixel via the count == 0 path) or to a wrong divisor (silently
incorrect pixels).

Compute the area in u64 via source_region_area() and thread it through as
u64, removing the re-cast at the divide site. A 65536x65536 source mapped
to 1x1 now yields the exact 2^32 divisor instead of overflowing.

Closes #76

+++ src/resize.rs
@@ -203,7 +203,7 @@ fn downscale_half_alpha(src: &Raster) -> Result<Raster, RasterError> {
 /// may hold more than `u32::MAX` samples, so the product is computed in `u64`.
 #[inline]
 fn source_region_area(sx0: u32, sx1: u32, sy0: u32, sy1: u32) -> u64 {
-    ((sx1 - sx0) * (sy1 - sy0)) as u64
+    (sx1 - sx0) as u64 * (sy1 - sy0) as u64
 }
 
 /// Downscale a raster to arbitrary dimensions using simple bilinear-ish area averaging.
